Convert isUnitlessNumber lookup table to memoized RegExp test
