+++ package.json
@@ -65,6 +65,7 @@
     "@types/react": "^16.4.6",
     "babel-core": "7.0.0-bridge.0",
     "babel-jest": "^23.4.2",
+    "babel-plugin-codegen": "^3.0.0",
     "babel-plugin-transform-react-remove-prop-types": "0.4.13",
     "enzyme": "^3.4.4",
     "enzyme-adapter-react-16": "^1.2.0",
@@ -93,6 +94,7 @@
   },
   "dependencies": {
     "@babel/runtime": "7.0.0-beta.51",
+    "@emotion/memoize": "^0.6.5",
     "memoize-one": "^4.0.0"
   },
   "jest": {
